+++ packages/pg/test/unit/client/stream-and-query-error-interaction-tests.js
@@ -1,12 +1,12 @@
 'use strict'
 const helper = require('./test-helper')
-var Connection = require('../../../lib/connection')
-var Client = require('../../../lib/client')
+const Connection = require('../../../lib/connection')
+const Client = require('../../../lib/client')
 const assert = require('assert')
 const suite = new helper.Suite()
 
 suite.test('emits end when not in query', function () {
-  var stream = new (require('events').EventEmitter)()
+  const stream = new (require('events').EventEmitter)()
   stream.setNoDelay = () => {}
   stream.connect = function () {
     // NOOP
@@ -15,7 +15,7 @@ suite.test('emits end when not in query', function () {
     // NOOP
   }
 
-  var client = new Client({ connection: new Connection({ stream: stream }) })
+  const client = new Client({ connection: new Connection({ stream: stream }) })
   client.connect(
     assert.calls(function () {
       client.query(
@@ -31,8 +31,6 @@ suite.test('emits end when not in query', function () {
   client.connection.emit('connect')
   process.nextTick(function () {
     client.connection.emit('readyForQuery')
-    assert.equal(client.queryQueue.length, 0)
-    assert(client.activeQuery, 'client should have issued query')
     process.nextTick(function () {
       stream.emit('close')
     })